Re-organise code
