+++ src/components/actions/getApplicationData.ts
@@ -62,7 +62,9 @@ export const getApplicationData = async (input: {
     webHostUrl: process.env.WEB_HOST,
   }
 
-  const sectionCodes = (await DBConnect.getApplicationSections(applicationId)).map(({code} : { code: string }) => code)
+  const sectionCodes = (await DBConnect.getApplicationSections(applicationId)).map(
+    ({ code }: { code: string }) => code
+  )
 
   return {
     action_payload: input?.payload,
@@ -71,6 +73,6 @@ export const getApplicationData = async (input: {
     responses: responseData,
     reviewData,
     environmentData,
-    sectionCodes
+    sectionCodes,
   }
 }
