+++ src/server.ts
@@ -20,7 +20,7 @@ import { routeGeneratePDF } from './components/files/documentGenerate'
 import { saveFiles, getFilePath, filesFolder } from './components/files/fileHandler'
 import { createDefaultDataFolders } from './components/files/createDefaultFolders'
 import { getAppEntryPointDir, objectKeysToSnakeCase } from './components/utilityFunctions'
-import { routeRunAction, routeGetApplicationData } from './components/actions/runAction'
+import { routeRunAction, routeGetApplicationData } from './components/actions'
 import config from './config'
 import lookupTableRoutes from './lookup-table/routes'
 import snapshotRoutes from './components/snapshots/routes'
